+++ R/legend.R
@@ -38,6 +38,14 @@
 #'
 #' in pixels
 #'
+#' @param group
+#'
+#' group name of a leaflet layer group
+#'
+#' @param className
+#'
+#' extra CSS class to append to the control, space separated
+#'
 #' @param ...
 #'
 #' arguments to pass to \link[leaflet]{addControl}
@@ -68,13 +76,37 @@
 #' )
 #'
 #' leaflet(data = quakes1) %>% addTiles() %>%
-#'   addMarkers(~long, ~lat, icon = leafIcons) %>%
+#'   addMarkers(~long, ~lat, icon = leafIcons, group = 'Quake Leaves') %>%
 #'   addLegendImage(images = c("http://leafletjs.com/examples/custom-icons/leaf-green.png",
 #'                             "http://leafletjs.com/examples/custom-icons/leaf-red.png"),
 #'                  labels = c('Green', 'Red'),width = 38, height = 95,
 #'                  title = htmltools::tags$div('Leaf',
 #'                  style = 'font-size: 24px; text-align: center;'),
-#'                  position = 'topright')
+#'                  position = 'topright',
+#'                  group = 'Quake Leaves') %>%
+#'   addLayersControl(overlayGroups = c('Quake Leaves'), position = 'bottomright')
+#'
+#'  # use raster images with size encodings
+#'  height <- sizeNumeric(quakes$depth, baseSize = 40)
+#'  width <- height * 38 / 95
+#'  symbols <- icons(
+#'    iconUrl = 'http://leafletjs.com/examples/custom-icons/leaf-green.png',
+#'    iconWidth = width,
+#'    iconHeight = height)
+#'  probs <- c(.2, .4, .6, .8)
+#'  leaflet(quakes) %>%
+#'    addTiles() %>%
+#'    addMarkers(icon = symbols,
+#'               lat = ~lat, lng = ~long) %>%
+#'    addLegendImage(
+#'      images = rep("http://leafletjs.com/examples/custom-icons/leaf-green.png", 4),
+#'      labels = round(quantile(height, probs = probs), 0),
+#'      width = quantile(height, probs = probs) * 38 / 95,
+#'      height = quantile(height, probs = probs),
+#'      title = htmltools::tags$div(
+#'        'Leaf',
+#'        style = 'font-size: 24px; text-align: center; margin-bottom: 5px;'),
+#'      position = 'topright', orientation = 'vertical')
 addLegendImage <- function(map,
                            images,
                            labels,
@@ -83,6 +115,8 @@ addLegendImage <- function(map,
                            orientation = c('vertical', 'horizontal'),
                            width = 20,
                            height = 20,
+                           group = NULL,
+                           className = 'info legend leaflet-control',
                            ...) {
   stopifnot(length(images) == length(labels))
   orientation <- match.arg(orientation)
@@ -90,6 +124,9 @@ addLegendImage <- function(map,
     htmlTag <- htmltools::tags$div
   } else {
     htmlTag <- htmltools::tags$span
+  }
+  if ( inherits(images, 'svgURI') ) {
+    images <- list(images)
   }
   htmlElements <- Map(
     img = images,
@@ -131,7 +168,7 @@ addLegendImage <- function(map,
     htmlElements <-
       append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))), after = 0)
   }
-  leaflet::addControl(map, html = htmltools::tagList(htmlElements), ...)
+  leaflegendAddControl(map, html = htmltools::tagList(htmlElements), className = className, group = group, ...)
 }
 
 #' Create an SVG tag for the symbol
@@ -433,6 +470,14 @@ makeSymbolIcons <- function(shape = c('rect',
 #'
 #' fill opacity of the legend items
 #'
+#' @param group
+#'
+#' group name of a leaflet layer group
+#'
+#' @param className
+#'
+#' extra CSS class to append to the control, space separated
+#'
 #' @param ...
 #'
 #' arguments to pass to \link[leaflet]{addControl}
@@ -563,6 +608,37 @@ makeSymbolIcons <- function(shape = c('rect',
 #'   addLegend(pal = binPal,
 #'             values = quakes$mag,
 #'             title = 'addLegend')
+#'
+#' # Group Layer Control
+#' # Works with baseGroups and overlayGroups
+#'
+#' leaflet() %>%
+#'   addTiles() %>%
+#'   addLegendNumeric(
+#'     pal = numPal,
+#'     values = quakes$depth,
+#'     position = 'topright',
+#'     title = 'addLegendNumeric',
+#'     group = 'Numeric Data'
+#'   ) %>%
+#'   addLegendQuantile(
+#'     pal = quantPal,
+#'     values = quakes$mag,
+#'     position = 'topright',
+#'     title = 'addLegendQuantile',
+#'     group = 'Quantile'
+#'   ) %>%
+#'   addLegendBin(
+#'     pal = binPal,
+#'     values = quakes$mag,
+#'     position = 'bottomleft',
+#'     title = 'addLegendBin',
+#'     group = 'Bin'
+#'   ) %>%
+#'   addLayersControl(
+#'     baseGroups = c('Numeric Data', 'Quantile'),  overlayGroups = c('Bin'),
+#'     position = 'bottomright'
+#'   )
 addLegendNumeric <- function(map,
                              pal,
                              values,
@@ -578,8 +654,11 @@ addLegendNumeric <- function(map,
                              tickWidth = 1,
                              decreasing = FALSE,
                              fillOpacity = 1,
+                             group = NULL,
+                             className = 'info legend leaflet-control',
                              ...) {
   stopifnot( attr(pal, 'colorType') == 'numeric' )
+  shape <- match.arg(shape)
   rng <- range(values, na.rm = TRUE)
   breaks <- pretty(values, bins)
   if ( breaks[1] < rng[1] ) {
@@ -708,7 +787,7 @@ addLegendNumeric <- function(map,
     htmlElements <-
       append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))), after = 0)
   }
-  leaflet::addControl(map, html = htmltools::tagList(htmlElements), ...)
+  leaflegendAddControl(map, html = htmltools::tagList(htmlElements), className = className, group = group, ...)
 }
 
 #' @export
@@ -727,6 +806,8 @@ addLegendQuantile <- function(map,
                               numberFormat = function(x) {prettyNum(x, big.mark = ',', scientific = FALSE, digits = 1)},
                               opacity = 1,
                               fillOpacity = opacity,
+                              group = NULL,
+                              className = 'info legend leaflet-control',
                               ...) {
   stopifnot( attr(pal, 'colorType') == 'quantile' )
   shape <- match.arg(shape)
@@ -766,7 +847,7 @@ addLegendQuantile <- function(map,
     htmlElements <-
       append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))), after = 0)
   }
-  leaflet::addControl(map, html = htmltools::tagList(htmlElements), ...)
+  leaflegendAddControl(map, html = htmltools::tagList(htmlElements), className = className, group = group, ...)
 }
 
 #' @export
@@ -784,6 +865,8 @@ addLegendBin <- function(map,
                          height = 24,
                          opacity = 1,
                          fillOpacity = opacity,
+                         group = NULL,
+                         className = 'info legend leaflet-control',
                          ...) {
   stopifnot( attr(pal, 'colorType') == 'bin' )
   shape <- match.arg(shape)
@@ -808,7 +891,7 @@ addLegendBin <- function(map,
     htmlElements <-
       append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))), after = 0)
   }
-  leaflet::addControl(map, html = htmltools::tagList(htmlElements), ...)
+  leaflegendAddControl(map, html = htmltools::tagList(htmlElements), className = className, group = group, ...)
 }
 
 #' @export
@@ -826,6 +909,8 @@ addLegendFactor <- function(map,
                             height = 24,
                             opacity = 1,
                             fillOpacity = opacity,
+                            group = NULL,
+                            className = 'info legend leaflet-control',
                             ...) {
   stopifnot( attr(pal, 'colorType') == 'factor' )
   shape <- match.arg(shape)
@@ -849,7 +934,7 @@ addLegendFactor <- function(map,
     htmlElements <-
       append(htmlElements, list(htmltools::div(htmltools::tags$strong(title))), after = 0)
   }
-  leaflet::addControl(map, html = htmltools::tagList(htmlElements), ...)
+  leaflegendAddControl(map, html = htmltools::tagList(htmlElements), className = className, group = group, ...)
 }
 
 #' Add a legend that for the sizing of symbols
@@ -864,7 +949,9 @@ addLegendFactor <- function(map,
 #'
 #' @param values
 #'
-#' the values used to generate colors from the palette function
+#' the values used to generate sizes and if colorValues is not specified and
+#' pal is given, then the values are used to generate  colors from the palette
+#' function
 #'
 #' @param title
 #'
@@ -912,6 +999,22 @@ addLegendFactor <- function(map,
 #' @param strokeWidth
 #'
 #' width of symbol outline
+#'
+#' @param colorValues
+#'
+#' the values used to generate color from the palette function
+#'
+#' @param numberFormat
+#'
+#' formatting functions for numbers that are displayed e.g. format, prettyNum
+#'
+#' @param group
+#'
+#' group name of a leaflet layer group
+#'
+#' @param className
+#'
+#' extra CSS class to append to the control, space separated
 #'
 #' @param ...
 #'
@@ -987,6 +1090,21 @@ addLegendFactor <- function(map,
 #'     color = 'black',
 #'     baseSize = 20,
 #'     breaks = 5)
+#'
+#' # Group layers control
+#' leaflet() %>%
+#'   addTiles() %>%
+#'     addLegendSize(
+#'       values = quakes$depth,
+#'       pal = numPal,
+#'       title = 'Depth',
+#'       labelStyle = 'margin: auto;',
+#'       shape = c('triangle'),
+#'       orientation = c('vertical', 'horizontal'),
+#'       opacity = .7,
+#'       breaks = 5,
+#'       group = 'Depth') %>%
+#'     addLayersControl(overlayGroups = c('Depth'))
 addLegendSize <- function(map,
                           pal,
                           values,
@@ -1001,6 +1119,9 @@ addLegendSize <- function(map,
                           fillOpacity = opacity,
                           breaks = 5,
                           baseSize = 10,
+                          numberFormat = function(x) {prettyNum(x, big.mark = ',', scientific = FALSE, digits = 1)},
+                          group = NULL,
+                          className = 'info legend leaflet-control',
                           ...) {
   shape <- match.arg(shape)
   sizes <- sizeBreaks(values, breaks, baseSize)
@@ -1011,7 +1132,11 @@ addLegendSize <- function(map,
     colors <- color
   }
   if ( missing(fillColor) ) {
-    fillColors <- pal(as.numeric(names(sizes)))
+    if ( !missing(pal) ) {
+      fillColors <- pal(as.numeric(names(sizes)))
+    } else {
+      fillColors <- colors
+    }
   } else {
     stopifnot(length(fillColor) == 1 || length(fillColor) == length(breaks))
     fillColors <- fillColor
@@ -1025,9 +1150,9 @@ addLegendSize <- function(map,
                  opacity = opacity,
                  fillOpacity = fillOpacity,
                  `stroke-width` = strokeWidth)
-  addLegendImage(map, images = symbols, labels = names(sizes),
+  addLegendImage(map, images = symbols, labels = numberFormat(as.numeric(names(sizes))),
                  title = title, labelStyle = labelStyle,
-                 orientation = orientation, width = sizes, height = sizes, ...)
+                 orientation = orientation, width = sizes, height = sizes, group = group, className = className, ...)
 
 }
 
@@ -1057,7 +1182,8 @@ makeSizeIcons <- function(values,
                                     'cross', 'diamond', 'star', 'stadium'),
                           pal,
                           color,
-                          fillColor = color,
+                          colorValues,
+                          fillColor,
                           opacity,
                           fillOpacity = opacity,
                           strokeWidth = 1,
@@ -1066,17 +1192,30 @@ makeSizeIcons <- function(values,
                           ) {
   shape <- match.arg(shape)
   if ( missing(color) ) {
-    colors <- pal(values)
+    if ( missing(colorValues) ) {
+      colors <- pal(values)
+    } else {
+      colors <- pal(colorValues)
+    }
   } else {
     stopifnot(length(color) == 1 || length(color) == length(values))
     colors <- color
   }
   if ( missing(fillColor) ) {
-    fillColors <- pal(values)
+    if ( !missing(pal) ) {
+      if ( missing(colorValues) ) {
+        fillColors <- pal(values)
+      } else {
+        fillColors <- pal(colorValues)
+      }
+    } else {
+      fillColors <- colors
+    }
   } else {
     stopifnot(length(fillColor) == 1 || length(fillColor) == length(values))
     fillColors <- fillColor
   }
+
   sizes <- sizeNumeric(values, baseSize)
   makeSymbolIcons(
     shape = shape,
@@ -1091,3 +1230,38 @@ makeSizeIcons <- function(values,
   )
 }
 
+leaflegendAddControl <- function(map,
+                                 html,
+                                 className,
+                                 group,
+                                 ...) {
+
+  if ( !is.null(group) ) {
+    leafLegendClassName = paste('leaflegend-group', gsub(' ', '', group), sep = '-')
+    className <- paste(className, leafLegendClassName)
+
+    lf <- leaflet::addControl(map, html = html, className = className, ...)
+    htmlwidgets::onRender(lf, "
+                function(el, x) {
+                  var updateLeafLegend = function() {
+                    var controlGroups = document.querySelectorAll('input.leaflet-control-layers-selector');
+                    controlGroups.forEach(g => {
+                      var groupName = g.nextSibling.innerText.substr(1);
+                      var className = 'leaflegend-group-' + groupName.replace(' ', '');
+                      var checked = g.checked;
+                      document.querySelectorAll('.legend.' + className).forEach(l => {
+                        l.hidden = !checked;
+                      })
+                    })
+                  }
+
+                  updateLeafLegend();
+                  this.on('baselayerchange', el => updateLeafLegend())
+                  this.on('overlayadd', el => updateLeafLegend());
+                  this.on('overlayremove', el => updateLeafLegend());
+                }
+                        ")
+  } else {
+    leaflet::addControl(map, html = html, className = className, ...)
+  }
+}